apsdbcli merged into apscli, test adapted accordingly
