+++ apg-patch-service-cmdclient/src/main/groovy/com/apgsga/patch/service/client/config/PliConfig.java
@@ -20,16 +20,10 @@ public class PliConfig {
 	
 	@Configuration
     @Profile("default")
-	@PropertySource({ "${appPropertiesFile}" })
+	@PropertySource({ "${appPropertiesFile}","${opsPropertiesFile}" })
     static class DefaultCliConfig
     { }
 	
-	@Configuration
-    @Profile("dbcli")
-	@PropertySource({ "${appPropertiesFile}", "${opsPropertiesFile}" })
-    static class DbCliConfig
-    { }
-
 	@Bean
 	public ConfigObject configObject() {
 		ConfigSlurper sl = new ConfigSlurper();
