+++ apg-patch-service-cmdclient/src/test/resources/config/app-test.properties
@@ -1,13 +1,12 @@
-revision.file.path=src/test/resources/Revisions.json
 host.default=localhost:9020
 config.dir=src/test/resources/config
-onclone.delete.artifact.dryrun=true
 target.system.mapping.file.name=TargetSystemMappings.json
-postclone.list.patch.filepath.template=src/test/resources/patchToBeReinstalled
-artifactory.dbpatch.repo.name=dbpatch
-artifactory.release.repo.name=releases
-patchRepoName=yumpatchrepo
-mavenRepoUser=dev
+artifactory.dbpatch.repo.name=dbpatch-test
+artifactory.release.repo.name=releases-test
+patchRepoName=yumpatchrepo-test
+mavenRepoUser=dev-test
 mavenRepoBaseUrl=https://artifactory4t4apgsga.jfrog.io/artifactory4t4apgsga
-mavenRepoName=repo
+mavenRepoName=repo-test
 json.db.location=src/test/resources
+# JHE (26.06.2020) : will probably be removed, but currently still needed by "lpac" option
+postclone.list.patch.filepath.template=src/test/resources/patchToBeReinstalled
