+++ apg-patch-service-cmdclient/src/test/resources/config/ops-test.properties
@@ -1,3 +1,3 @@
-db.url=jdbc:oracle:thin:@chei211.apgsga.ch:1521:chei211
+db.url=jdbc:oracle:thin:@chei212.apgsga.ch:1521:chei212
 db.user=cm_installer
 db.passwd=cm_installer_pass
