+++ apg-patch-service-cmdclient/src/test/resources/config/server-test.properties
@@ -7,6 +7,6 @@ json.db.work.location=build/testdbwork
 config.common.location=src/test/resources/config
 config.common.targetSystemFile=TargetSystemMappings.json
 patch.action.script.location=classpath:groovyScript/executePatchAction.groovy
-mavenRepoUser=dev
+mavenRepoUser=dev-test
 mavenRepoBaseUrl=https://artifactory4t4apgsga.jfrog.io/artifactory4t4apgsga
-mavenRepoName=repo
+mavenRepoName=repo-test
