+++ apg-patch-service-server/db/Patch5402.json
@@ -1,1 +1,1 @@
-{"patchNummer":"5402","dbPatchBranch":"Patch_0900C1_5402","prodBranch":"HEAD","patchTag":"Patch_0900C1_5402_2","developerBranch":"","tagNr":2,"installationTarget":"CHEI212","targetToState":null,"runningNr":null,"dbObjects":[{"fileName":"FileName1","filePath":"FilePath1","moduleName":null}],"dockerServices":[],"services":[{"serviceName":"It21Ui","microServiceBranch":null,"baseVersionNumber":"XXXX","revisionMnemoPart":null,"installOnEmptyModules":false,"revision":null,"lastRevision":null,"mavenArtifacts":[],"mavenArtifactsAsVcsPath":[],"mavenArtifactsToBuild":[],"installJadasAndGui":false}],"installDockerServices":false,"lastPipelineTask":"","currentTarget":null,"currentPipelineTask":null,"logText":null,"dbObjectsAsVcsPath":["FilePath1/FileName1"],"mavenArtifacts":[]}
+{"patchNummer":"5402","dbPatchBranch":"Patch_0900C1_5402","prodBranch":"HEAD","patchTag":"Patch_0900C1_5402_2","developerBranch":"","tagNr":2,"installationTarget":"CHEI212","targetToState":null,"runningNr":null,"dbObjects":[{"fileName":"FileName1","filePath":"FilePath1","moduleName":null}],"dockerServices":[],"services":[{"serviceName":"It21Ui","microServiceBranch":null,"baseVersionNumber":"XXXX","revisionMnemoPart":null,"installOnEmptyModules":false,"revision":null,"lastRevision":null,"mavenArtifacts":[],"mavenArtifactsToBuild":[],"installJadasAndGui":false,"mavenArtifactsAsVcsPath":[]}],"installDockerServices":false,"lastPipelineTask":"","currentTarget":null,"currentPipelineTask":null,"logText":null,"dbObjectsAsVcsPath":["FilePath1/FileName1"],"mavenArtifacts":[]}
